% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_qtl2_to_qtl.R
\name{scan_qtl2_to_qtl}
\alias{scan_qtl2_to_qtl}
\title{Convert scan1 results to the scanone format}
\usage{
scan_qtl2_to_qtl(scan_output)
}
\arguments{
\item{scan_output}{Matrix of LOD scores, as calculated by
\code{qtl2::scan1}.}
}
\value{
A data frame with class \code{"scanone"}, containing
chromosome and position columns followed by the LOD scores in
\code{scan_output}.
}
\description{
Convert the results of \code{qtl2::scan1}
to the form used by the R/qtl function
\code{qtl::scanone}.
}
\examples{
library(qtl2geno)
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))
probs <- calc_genoprob(iron, step=1, error_prob=0.002)
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)
library(qtl2scan)
out <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar)

out_rev <- scan_qtl2_to_qtl(out)

}
