% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2csv.R
\name{write2csv}
\alias{write2csv}
\title{Write a data frame to a CSV file}
\usage{
write2csv(df, filename, comment = "", sep = ",", comment.char = "#")
}
\arguments{
\item{df}{A data frame (or matrix)}

\item{filename}{File name to write}

\item{comment}{Comment to place on the first line}

\item{sep}{Field separator}

\item{comment.char}{Character to use to initiate the comment lines}
}
\description{
Write a data frame to a CSV file in a special form, with info about the number of rows and columns.
}
\details{
If the file already exists, the function will refuse to write over it.

The file will include comments at the top, using \code{#} as a
comment character, including the number of rows (not including the
header) and the number of columns.

Row names are not included.
}
\examples{
nr <- 10
nc <- 5
x <- data.frame(id=paste0("ind", 1:nr),
                matrix(rnorm(nr*nc), ncol=nc))
colnames(x)[1:nc + 1] <- paste0("col", 1:nc)
\dontrun{
write2csv(x, "/tmp/tmpfile.csv", "A file created by write2csv")}
}

