% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_geno.R
\name{encode_geno}
\alias{encode_geno}
\title{Encode a matrix of genotypes using a set of allele codes}
\usage{
encode_geno(geno, allele_codes, output_codes = c("-", "A", "H", "B"))
}
\arguments{
\item{geno}{Character matrix of genotypes}

\item{allele_codes}{Two-column matrix of alleles}

\item{output_codes}{Vector of length four, with missing, AA, AB, BB codes}
}
\value{
Matrix of same dimensions as \code{geno}, but with values in \code{output_codes}.
}
\description{
Encode a matrix of genotypes using a set of allele codes.
}
\examples{
geno <- rbind(c("C", "G", "C",  "GG", "CG"),
              c("A", "A", "AT", "TA", "TT"),
              c("T", "G", NA,   "GT", "TT"))
codes <- rbind(c("C", "G"), c("A", "T"), c("T", "G"))
encode_geno(geno, codes)

}
\seealso{
\code{\link{find_consensus_geno}}, \code{\link{find_unique_geno}}
}
