% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_df_to_list.R
\name{map_df_to_list}
\alias{map_df_to_list}
\title{Marker map data frame to list}
\usage{
map_df_to_list(map, chr_column = "chr", pos_column = "cM",
  marker_column = "marker")
}
\arguments{
\item{map}{Data frame with marker map}

\item{chr_column}{Name of the column in \code{map} that contains the chromosome IDs.}

\item{pos_column}{Name of the column in \code{map} that contains the marker positions.}

\item{marker_column}{Name of the column in \code{map} that contains
the marker names. If NULL, use the row names.}
}
\value{
A list of vectors of marker positions, one component per chromosome
}
\description{
Convert a marker map organized as data frame to a list
}

