% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs_doqtl_to_qtl2.R
\name{probs_doqtl_to_qtl2}
\alias{probs_doqtl_to_qtl2}
\title{Convert DOQTL genotype probabilities to R/qtl2 format}
\usage{
probs_doqtl_to_qtl2(probs, map, chr_column = "chr", pos_column = "cM",
  marker_column = "marker")
}
\arguments{
\item{probs}{3d array of genotype probabilities as calculated from DOQTL}

\item{map}{Data frame with marker map}

\item{chr_column}{Name of the column in \code{map} that contains the chromosome IDs.}

\item{pos_column}{Name of the column in \code{map} that contains the marker positions.}

\item{marker_column}{Name of the column in \code{map} that contains
the marker names. If NULL, use the row names.}
}
\value{
An object of the for produced by the R/qtl2geno function \code{calc_genoprob}.
}
\description{
Convert DOQTL genotype probabilities to R/qtl2 format
}

