% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl2convert-package.R
\docType{package}
\name{qtl2convert-package}
\alias{qtl2convert}
\alias{qtl2convert-package}
\title{qtl2convert: Convert Data among QTL Mapping Packages}
\description{
Functions to convert data structures among the 'qtl2', 'qtl', and 'DOQTL' packages for mapping quantitative trait loci (QTL).
}
\section{Vignettes}{

\itemize{
\item \href{https://kbroman.org/qtl2/assets/vignettes/user_guide.html}{user guide}
\item \href{https://kbroman.org/qtl2/assets/vignettes/input_files.html}{input file formats}
\item \href{https://kbroman.org/qtl2/pages/prep_do_data.html}{preparing DO mouse data for R/qtl2}
\item \href{https://kbroman.org/qtl2/assets/vignettes/rqtl_diff.html}{differences between R/qtl and R/qtl2}
\item \href{https://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{developer guide}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{http://kbroman.org/qtl2}
  \item \url{https://github.com/rqtl/qtl2convert}
  \item Report bugs at \url{https://github.com/rqtl/qtl2convert/issues}
}

}
\author{
\strong{Maintainer}: Karl W Broman \email{broman@wisc.edu} (\href{https://orcid.org/0000-0002-4914-6671}{ORCID})

}
\keyword{internal}
